Re-factor code, add SNMP support
The none packet-broker specific code of configure.py has been moved to
a separate module bfrt.py, configure.py is renamed to config.py.

The Config class is moved to a separate module config.py to make the
code more readable.

Add code to interface with the snabb-snmp-subagent to provide interface
MIBs for all interfaces.

Add init.d service to start the subagent on startup.