Handle the case when all members of a group are down
If all members of an action selector are down, the entry can only be
modified if no references to it exist. We must add or delete the
entries in the forwarding table accordingly.

When a forwarding entry is removed, the table must mark packets
destined for that group to be dropped.