Add ingress-ports, non-ip property to flow-mirror configurations
The optional ingress-ports property is a list of ingress ports to
which to apply the flow-mirror rule. By default, the list is empty and
the rule is applied to all ports.

The optional boolean property "non-ip" mirrors non-ip traffic if set to
true.  In this case, all other match fields are ignored.